requirements-dev.txt
boto3
flask-lambda-support
pynamodb
pylint
black
isort